Remove duplicate KeyMetadata class and move KeyGen to the ssh module
